requirements.txt
pydantic==2.10.1
chainlit==2.0.3
langchain==0.3.14
langgraph==0.2.62
prettytable==3.12.0
langchain-openai==0.3.0
langchain-nvidia-ai-endpoints==0.3.5
PyYAML==6.0.2
fastapi==0.115.6
uvicorn==0.25.0
tiktoken==0.8.0
pandas==2.2.3
markdown-pdf==1.3.1
cvss==3.3
aiosqlite==0.20.0
SQLAlchemy==2.0.36
greenlet==3.1.1
psycopg2-binary>=2.9.0
aiohttp==3.11.7
asyncpg==0.30.0
sqlparse==0.5.3
prompt_toolkit==3.0.50
